Paragraph 12.2.2 Knuth-Morris-Pratt pattern matching.
