+++ src/test/java/nl/papendorp/chapter12/kmp/KnuthMorrisPratSuite.java
@@ -1,0 +1,13 @@
+package nl.papendorp.chapter12.kmp;
+
+import nl.papendorp.chapter12.PatternMatcher;
+import nl.papendorp.chapter12.PatternMatchingSuite;
+
+public class KnuthMorrisPratSuite extends PatternMatchingSuite
+{
+	@Override
+	protected PatternMatcher sut()
+	{
+		return new KnuthMorrisPratt();
+	}
+}
